full grid field simulation parameter sweeps (150 pA)
These simulations test whether the presence/absence of noise helps in the path
integration process of the network. There is a simulation of the full animal
movement (here in particular 600s), with the initialisation place cells and
velocity place cells, currently for one trial. This trial will be evaluated and
subsequently more trials will be run.

- Saving full population of E and spikes of the first 10 I neurons
- Every simulation should have its bumpCurrentSlope parameter correctly set up
  from the estimated data
- if bumpCurrentSlope does not exist in a simulation, it is aborted
- A -notify option added to qsub and when the simulation receives a signal
  (SIGUSR2) it attempts to store all the data before SIGKILL arrives

+++ grid_cell_model/simulations/007_noise/simulation_grids.py
@@ -18,24 +18,41 @@
 #       You should have received a copy of the GNU General Public License
 #       along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #
+import numpy as np
+from os.path            import exists
 from numpy.random       import choice
 
 from models.parameters  import getOptParser
 from models.gc_net_nest import BasicGridCellNetwork
 from data_storage       import DataStorage
+from nest.hl_api        import NESTError
 
 
 parser          = getOptParser()
 parser.add_option("--velON", type="int", help="Velocity input ON?")
 (options, args) = parser.parse_args()
 
+
+# Do nothing when bumpCurrentSlope is NaN
+if (np.isnan(options.bumpCurrentSlope)):
+    print('bumpCurrentSlope is NaN. Not performing the simulation')
+    exit(0)
+else:
+    print("bumpCurrentSlope: {0}".format(options.bumpCurrentSlope))
+
+
 stateMonParams = {
         'start' : options.time - options.stateMonDur
 }
+nrec_spikes_e = None # all neurons
+nrec_spikes_i = 10
 
 
 output_fname = "{0}/{1}job{2:05}_output.h5".format(options.output_dir,
         options.fileNamePrefix, options.job_num)
+if (exists(output_fname)):
+    print('Output file exists, doing nothing')
+    exit(0)
 d = DataStorage.open(output_fname, 'w')
 
 out = []
@@ -44,13 +61,18 @@ overalT = 0.
 for trial_idx in range(options.ntrials):
     print("\n\t\tStarting trial no. {0}\n".format(trial_idx))
     ei_net = BasicGridCellNetwork(options, simulationOpts=None,
+            nrec_spikes=(nrec_spikes_e, nrec_spikes_i),
             stateRecParams=(stateMonParams, stateMonParams))
 
     if (options.velON == 1):
         ei_net.setVelocityCurrentInput_e()
     ei_net.setPlaceCells()
     
-    ei_net.simulate(options.time, printTime=options.printTime)
+    try:
+        ei_net.simulate(options.time, printTime=options.printTime)
+    except NESTError as e:
+        print("Simulation interrupted. Message: {0}".format(str(e)))
+        print("Trying to save the simulated data if possible...")
     ei_net.endSimulation()
     out.append(ei_net.getAllData())
     constrT, simT, totalT = ei_net.printTimes()
